remove and correct warnings in classes

+++ projectticket/projectticket/Hall.cpp
@@ -7,7 +7,7 @@ Hall::Hall()
     this->numberSeatsPerRows = 0;
 }
 
-Hall::Hall(int numberRows, int numberSeatsPerRows)
+Hall::Hall(size_t numberRows, size_t numberSeatsPerRows)
 {
 
     setNumberRows(numberRows);
@@ -18,12 +18,12 @@ void Hall::setHallNumber(const size_t hallNumber)
     this->hallNumber = hallNumber;
 }
 
-void Hall::setNumberRows(const int numberRows)
+void Hall::setNumberRows(const size_t numberRows)
 {
     this->numberRows = numberRows;
 }
 
-void Hall::setNumberColumns(const int numberSeatsPerRows)
+void Hall::setNumberColumns(const size_t numberSeatsPerRows)
 {
     this->numberSeatsPerRows = numberSeatsPerRows;
 }
@@ -33,12 +33,12 @@ size_t Hall::getHallNumber() const
     return this->hallNumber;
 }
 
-int Hall::getNumberRows() const
+size_t Hall::getNumberRows() const
 {
     return this->numberRows;
 }
 
-int Hall::getNumberColumns() const
+size_t Hall::getNumberColumns() const
 {
     return this->numberSeatsPerRows;
 }
@@ -50,7 +50,7 @@ void Hall::book(const Date& date)
 
 bool Hall::isFree(const Date& date) const
 {
-    for (int i = 0; i < bookedDates.getSize(); i++)
+    for (size_t i = 0; i < bookedDates.getSize(); i++)
     {
         if (bookedDates[i] == date)
         {
