+++ projectticket/projectticket/Hall.h
@@ -8,20 +8,20 @@ class Hall {
 private:
 
     size_t hallNumber;
-    int numberRows;
-    int numberSeatsPerRows;
+    size_t numberRows;
+    size_t numberSeatsPerRows;
     Vector<Date> bookedDates;
 public:
     Hall();
-    Hall(int numberRows, int numberSeatsPerRows);
+    Hall(size_t numberRows, size_t numberSeatsPerRows);
 
     void setHallNumber(const size_t hallNumber);
-    void setNumberRows(const int numberRows);
-    void setNumberColumns(const int numberSeatsPerRows);
+    void setNumberRows(const size_t numberRows);
+    void setNumberColumns(const size_t numberSeatsPerRows);
 
     size_t getHallNumber() const;
-    int getNumberRows() const;
-    int getNumberColumns() const;
+    size_t getNumberRows() const;
+    size_t getNumberColumns() const;
 
     void book(const Date& date);
     bool isFree(const Date& date) const;
